+++ src/orcaset/__init__.py
@@ -2,7 +2,8 @@
 # SPDX-License-Identifier: SSPL-1.0
 
 from orcaset.context import Context, CycleError, DepNode
-from orcaset.period import Period
+from orcaset.maybe import Maybe, Na, add_values, combine_values, isna
+from orcaset.period import Period, period_union
 from orcaset.queries import flow, level, overlapping, prorated, time_weighted
 from orcaset.rule import Demand, Rule, Step, fetch
 from orcaset.series import (
@@ -10,15 +11,16 @@ from orcaset.series import (
     GridSeries,
     Key,
     Keys,
+    MapNFn,
+    MapNSeries,
     MapSeries,
-    Maybe,
-    Na,
+    MergeKeysFn,
     ReduceFn,
     Replayable,
     SelectFn,
     Series,
+    SeriesSources,
     ValueFn,
-    isna,
 )
 from orcaset.yf import YF
 
@@ -32,8 +34,11 @@ __all__ = [
     "GridSeries",
     "Key",
     "Keys",
+    "MapNFn",
+    "MapNSeries",
     "MapSeries",
     "Maybe",
+    "MergeKeysFn",
     "Na",
     "Period",
     "ReduceFn",
@@ -41,13 +46,17 @@ __all__ = [
     "Rule",
     "SelectFn",
     "Series",
+    "SeriesSources",
     "Step",
     "ValueFn",
+    "add_values",
+    "combine_values",
     "fetch",
     "flow",
     "isna",
     "level",
     "overlapping",
+    "period_union",
     "prorated",
     "time_weighted",
 ]